+++ lib/src/nuvigator.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-import 'legacy_nurouter.dart' as legacy;
-import 'next/v1/nu_router.dart';
+import 'nu_router.dart';
 import 'screen_type.dart';
 import 'screen_types/material_screen_type.dart';
 import 'typings.dart';
@@ -11,18 +10,6 @@ enum DeepLinkPushMethod {
   Push,
   PushReplacement,
   PopAndPush,
-}
-
-NuvigatorState _tryToFindNuvigatorForRouter<T extends INuRouter>(
-    NuvigatorState nuvigatorState) {
-  if (nuvigatorState == null) return null;
-  final nuvigatorRouterForType = nuvigatorState.router.getRouter<T>();
-  if (nuvigatorRouterForType != null) return nuvigatorState;
-  if (nuvigatorState != nuvigatorState.parent &&
-      nuvigatorState.parent != null) {
-    return _tryToFindNuvigatorForRouter<T>(nuvigatorState.parent);
-  }
-  return null;
 }
 
 class NuvigatorStateTracker extends NavigatorObserver {
@@ -62,19 +49,24 @@ class NuvigatorStateTracker extends NavigatorObserver {
 }
 
 abstract class INuRouter {
+  /// When the [NuvigatorState] is initialized this method is called on the respective
+  /// [INuRouter] to register it's reference.
   void install(NuvigatorState nuvigator);
 
+  /// When the [NuvigatorState] disposes, this method is called on the respective
+  /// [INuRouter] to perform any cleanup work required.
   void dispose();
 
+  /// In case this [INuRouter] is not able to provide a [Route] for the requested
+  /// deepLink, and this property is different from null, it will be called to perform
+  /// a fallback execution
   HandleDeepLinkFn onDeepLinkNotFound;
 
-  @deprecated
-  T getRouter<T extends INuRouter>();
-
+  /// When a new deepLink is requested to be opened the [Nuvigator] this method is
+  /// called in the respective [INuRouter] to retrieve the [Route] to be presented
   Route<T> getRoute<T>({
     String deepLink,
     Object parameters,
-    bool fromLegacyRouteName = false,
     bool isFromNative = false,
     ScreenType fallbackScreenType,
     ScreenType overrideScreenType,
@@ -84,8 +76,7 @@ abstract class INuRouter {
 class _NuvigatorInner<T extends INuRouter> extends Navigator {
   _NuvigatorInner({
     @required this.router,
-    String initialRoute,
-    String initialDeepLink,
+    @required String initialDeepLink,
     Map<String, Object> initialArguments,
     Key key,
     List<NavigatorObserver> observers = const [],
@@ -95,23 +86,21 @@ class _NuvigatorInner<T extends INuRouter> extends Navigator {
     this.inheritableObservers = const [],
     this.shouldPopRoot = false,
   })  : assert(router != null),
-        assert((initialRoute == null) != (initialDeepLink == null)),
+        assert(initialDeepLink != null),
         super(
           observers: [
             HeroController(),
             ...observers,
           ],
           onGenerateInitialRoutes: (_, __) {
-            final deepLink = initialDeepLink ?? initialRoute;
             final r = router.getRoute<dynamic>(
-              deepLink: deepLink,
+              deepLink: initialDeepLink,
               parameters: initialArguments,
-              fromLegacyRouteName: initialDeepLink == null,
               fallbackScreenType: screenType,
             );
             if (r == null) {
               throw FlutterError(
-                  'No Route was found for the initialRoute provided: "$deepLink"'
+                  'No Route was found for the initialRoute provided: "$initialDeepLink"'
                   ' .Be sure that the provided initialRoute exists in this Router ($router).');
             }
             return [r];
@@ -120,7 +109,6 @@ class _NuvigatorInner<T extends INuRouter> extends Navigator {
             return router.getRoute<dynamic>(
               deepLink: settings.name,
               parameters: settings.arguments,
-              fromLegacyRouteName: true,
               fallbackScreenType: screenType,
             );
           },
@@ -174,8 +162,6 @@ class NuvigatorState<T extends INuRouter> extends NavigatorState
   Route get currentRoute => stateTracker.stack.last;
 
   INuRouter get rootRouter => rootNuvigator.router;
-
-  R getRouter<R extends INuRouter>() => router.getRouter<R>();
 
   List<ObserverBuilder> _collectObservers() {
     if (isNested) {
@@ -258,7 +244,6 @@ class NuvigatorState<T extends INuRouter> extends NavigatorState
       deepLink: routeName,
       parameters: parameters,
       isFromNative: false,
-      fromLegacyRouteName: true,
       fallbackScreenType: widget.screenType,
     );
   }
@@ -432,13 +417,10 @@ class NuvigatorState<T extends INuRouter> extends NavigatorState
 
   /// Prefer using [NuvigatorState.open], `.openDeepLink` does not support opening nested deepLinks
   Future<R> openDeepLink<R>(Uri deepLink, [dynamic arguments]) {
-    if (rootRouter is legacy.NuRouter) {
-      // ignore: avoid_as
-      return (rootRouter as legacy.NuRouter)
-          .openDeepLink<R>(deepLink, arguments, false);
-    } else {
-      return rootNuvigator.open(deepLink.toString(), parameters: arguments);
-    }
+    return rootNuvigator.open(
+      deepLink.toString(),
+      parameters: arguments,
+    );
   }
 
   /// Open the requested deepLink, if the current Nuvigator is not able to handle
@@ -460,7 +442,6 @@ class NuvigatorState<T extends INuRouter> extends NavigatorState
       deepLink: deepLink,
       parameters: parameters,
       isFromNative: isFromNative,
-      fromLegacyRouteName: false,
       fallbackScreenType: widget.screenType,
       overrideScreenType: screenType,
     );
@@ -513,33 +494,22 @@ class NuvigatorState<T extends INuRouter> extends NavigatorState
 }
 
 /// Creates a new Nuvigator. When using the Next API, several of those options
-/// are provided by the [NuRouter]. Providing them here will thrown an assertion
+/// are provided by the [INuRouter]. Providing them here will thrown an assertion
 /// error.
 @immutable
 class Nuvigator<T extends INuRouter> extends StatelessWidget {
-  Nuvigator({
+  const Nuvigator({
     @required this.router,
-    this.initialRoute,
-    this.initialDeepLink,
     this.initialArguments,
     Key key,
     this.observers = const [],
-    this.screenType,
     this.wrapper,
     this.debug = false,
     this.inheritableObservers = const [],
     this.shouldPopRoot = false,
     this.shouldRebuild,
   })  : _innerKey = key,
-        assert(router != null),
-        assert(() {
-          if (router is NuRouter) {
-            return initialDeepLink == null &&
-                initialRoute == null &&
-                screenType == null;
-          }
-          return true;
-        }());
+        assert(router != null);
 
   /// Creates a [Nuvigator] from a list of [NuRoute]
   static Nuvigator<NuRouterBuilder> routes({
@@ -571,20 +541,12 @@ class Nuvigator<T extends INuRouter> extends StatelessWidget {
   final T router;
   final bool debug;
   final bool shouldPopRoot;
-  final ScreenType screenType;
   final WrapperFn wrapper;
   final List<ObserverBuilder> inheritableObservers;
   final List<NavigatorObserver> observers;
   final Key _innerKey;
-  final String initialRoute;
-  final Uri initialDeepLink;
   final Map<String, Object> initialArguments;
   final ShouldRebuildFn shouldRebuild;
-
-  static NuvigatorState ofRouter<T extends INuRouter>(BuildContext context) {
-    final closestNuvigator = context.findAncestorStateOfType<NuvigatorState>();
-    return _tryToFindNuvigatorForRouter<T>(closestNuvigator);
-  }
 
   /// Fetches a [NuvigatorState] from the current BuildContext.
   static NuvigatorState<T> of<T extends INuRouter>(
@@ -594,20 +556,22 @@ class Nuvigator<T extends INuRouter> extends StatelessWidget {
   }) {
     if (rootNuvigator) {
       return context.findRootAncestorStateOfType<NuvigatorState<T>>();
+    } else {
+      final closestNuvigator =
+          context.findAncestorStateOfType<NuvigatorState<T>>();
+      if (closestNuvigator != null) return closestNuvigator;
+      assert(() {
+        if (!nullOk) {
+          throw FlutterError(
+              'Nuvigator operation requested with a context that does not include a Nuvigator.\n'
+              'The context used to push or pop routes from the Nuvigator must be that of a '
+              'widget that is a descendant of a Nuvigator widget.'
+              'Also check if the provided Router [T] type exists withing a the Nuvigator context.');
+        }
+        return true;
+      }());
+      return null;
     }
-    final nuvigatorState = ofRouter<T>(context);
-    if (nuvigatorState is NuvigatorState<T>) return nuvigatorState;
-    assert(() {
-      if (!nullOk) {
-        throw FlutterError(
-            'Nuvigator operation requested with a context that does not include a Nuvigator.\n'
-            'The context used to push or pop routes from the Nuvigator must be that of a '
-            'widget that is a descendant of a Nuvigator widget.'
-            'Also check if the provided Router [T] type exists withing a the Nuvigator context.');
-      }
-      return true;
-    }());
-    return null;
   }
 
   /// Helper method that allows passing a Nuvigator to a builder function
@@ -615,7 +579,8 @@ class Nuvigator<T extends INuRouter> extends StatelessWidget {
     return this;
   }
 
-  Widget _buildModuleRouter(BuildContext context) {
+  @override
+  Widget build(BuildContext context) {
     return NuRouterLoader(
       // ignore: avoid_as
       router: router as NuRouter,
@@ -632,30 +597,5 @@ class Nuvigator<T extends INuRouter> extends StatelessWidget {
         shouldPopRoot: shouldPopRoot,
       ),
     );
-  }
-
-  Widget _buildLegacyRouter(BuildContext context) {
-    return _NuvigatorInner<T>(
-      router: router,
-      debug: debug,
-      inheritableObservers: inheritableObservers,
-      observers: observers,
-      initialDeepLink: initialDeepLink?.toString(),
-      initialRoute: initialRoute,
-      screenType: screenType ?? materialScreenType,
-      key: _innerKey,
-      initialArguments: initialArguments,
-      wrapper: wrapper,
-      shouldPopRoot: shouldPopRoot,
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (router is NuRouter) {
-      return _buildModuleRouter(context);
-    } else {
-      return _buildLegacyRouter(context);
-    }
   }
 }