+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+
+## UNRELEASED (2.0.0)
+- [BREAKING] Remove Legacy API and Deprecated options related to it
+
 ## 1.8.0
 - [FIX] Make Nuvigator compatible with Flutter 2.10.5
 - [FIX] Make Nuvigator compatible with Flutter 3.0.2